finish first viriosn
